Add files via upload
Two structures are declared. One, describing an employee from a company, with the following fields: first name (first name, patronymic, last name), personal ID, work experience, no. workdays in the month, an array containing the wages for each day of the month, and the other structure include: the company name, employees array, and the number of employees. The program code contains the following instructions:
1. a function that enters data about an employee.
2. a function that calculates the salary of the employee for the month (sum of wages + length of service*1.5%* sum of wages - 25% for taxes).
3. displays the following information about the employee (name and salary).
In the main program, an appropriate menu is formed with the possibility of data entry, reference, and output.